# docs/v1/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API gateway",
        "title": "Web api gateway API",
        "contact": {
            "name": "ru.system.ru@gmail.com",
            "email": "ru.system.ru@gmail.com"
        },
        "version": "1.0"
    },
    "host": "91.243.71.100:8080",
    "basePath": "/api/web-api-gateway/v1",
    "paths": {
...
        "/warehouse": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Получение списка складов",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "warehouse"
                ],
                "summary": "Get warehouses",
                "operationId": "get-warehouses",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Warehouse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Создание склада",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "warehouse"
                ],
                "summary": "Create warehouse",
                "operationId": "create-warehouse",
                "parameters": [
                    {
                        "description": "Необходимо указать данные склада.",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.InputWarehouse"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.IdResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/warehouse/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Получение склада по id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "warehouse"
                ],
                "summary": "Get warehouse by id",
                "operationId": "get-warehouse",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Warehouse ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Warehouse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Обновление склада своей компании",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "warehouse"
                ],
                "summary": "Update warehouse",
                "operationId": "update-warehouse",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Warehouse ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Необходимо указать данные склада.",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.InputWarehouse"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Удаление склада своей компании",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "warehouse"
                ],
                "summary": "Delete warehouse",
                "operationId": "delete-warehouse",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Warehouse ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "domain.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "domain.IdResponse": {
            "type": "object",
            "properties": {
                "id": {}
            }
        },
...
        "domain.InputWarehouse": {
            "type": "object",
            "required": [
                "address",
                "name",
                "phone",
                "responsible_person"
            ],
            "properties": {
                "address": {
                    "description": "Адрес склада",
                    "type": "string",
                    "maxLength": 140,
                    "minLength": 5,
                    "example": "Адрес склада"
                },
                "country": {
                    "description": "Страна склада",
                    "type": "string",
                    "example": "Страна склада"
                },
                "current_occupancy": {
                    "description": "Текущая заполняемость склада",
                    "type": "integer",
                    "example": 50
                },
                "email": {
                    "description": "Электронная почта для связи",
                    "type": "string",
                    "example": "Электронная почта для связи"
                },
                "max_capacity": {
                    "description": "Максимальная вместимость склада",
                    "type": "integer",
                    "example": 100
                },
                "name": {
                    "description": "Название склада",
                    "type": "string",
                    "maxLength": 140,
                    "minLength": 5,
                    "example": "Название склада"
                },
                "other_fields": {
                    "description": "Дополнительные пользовательские поля",
                    "type": "object",
                    "additionalProperties": true
                },
                "phone": {
                    "description": "Контактный телефон склада",
                    "type": "string",
                    "minLength": 5,
                    "example": "Контактный телефон склада"
                },
                "responsible_person": {
                    "description": "Ответственное лицо за склад",
                    "type": "string",
                    "maxLength": 140,
                    "minLength": 5,
                    "example": "Ответственное лицо за склад"
                }
            }
        },
...
        "domain.Warehouse": {
            "type": "object",
            "properties": {
                "address": {
                    "description": "Адрес склада",
                    "type": "string"
                },
                "company_id": {
                    "description": "Уникальный идентификатор компании",
                    "type": "integer"
                },
                "country": {
                    "description": "Страна склада",
                    "type": "string"
                },
                "current_occupancy": {
                    "description": "Текущая заполняемость склада",
                    "type": "integer"
                },
                "email": {
                    "description": "Электронная почта для связи",
                    "type": "string"
                },
                "id": {
                    "description": "Уникальный идентификатор склада",
                    "type": "integer"
                },
                "max_capacity": {
                    "description": "Максимальная вместимость склада",
                    "type": "integer"
                },
                "name": {
                    "description": "Название склада",
                    "type": "string"
                },
                "other_fields": {
                    "description": "Дополнительные пользовательские поля",
                    "type": "object",
                    "additionalProperties": true
                },
                "phone": {
                    "description": "Контактный телефон склада",
                    "type": "string"
                },
                "responsible_person": {
                    "description": "Ответственное лицо за склад",
                    "type": "string"
                }
            }
        },
        "v1.response": {
            "type": "object",
            "properties": {
                "allow_registration": {
                    "type": "boolean"
                },
                "avatar": {
                    "type": "string"
                },
                "created_user_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_admin": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "sections": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "status": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}